+++ examples/montecarlo-cn0-target/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "montecarlo-cn0-target"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+montycarlo = "0.1.0"
+rand = "0.8"
